refactor: extract concurrent_structures to proximadb-concurrent foundation crate
Move src/infrastructure/concurrent_structures.rs (746 LOC, hermetic —
anyhow/dashmap/serde/bincode only, zero crate:: imports) into a new foundation
crate `proximadb-concurrent`. The root keeps a pure `pub use proximadb_concurrent::*`
re-export shim so every `crate::infrastructure::concurrent_structures::*` path
resolves unchanged — no caller edits.

First increment of the shared hermetic-leaf unlock in the root-crate
decomposition track. concurrent_structures is the most-depended-on hermetic
root-internal file (52 structural edges from the ORION graph engine and the
pgwire layer alone, plus dual_store_ivf and adaptive_structures). Extracting it
to a leaf removes a recurring root-internal dependency that would otherwise
force those cohorts into a cyclic root dependency when they are extracted next.

Pure relocation + re-export, no behavior change. Crate-level clippy allows
mirror the root crate (code was written under those suppressions).
