+++ crates/foundation/proximadb-concurrent/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "proximadb-concurrent"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+authors.workspace = true
+repository.workspace = true
+homepage.workspace = true
+description = "ProximaDB foundation-tier reusable concurrent data structures (ConcurrentStorage, AtomicMetrics, ConcurrentMapping, TypedStorage) shared by cache systems and index implementations"
+
+[lib]
+name = "proximadb_concurrent"
+path = "src/lib.rs"
+
+[dependencies]
+anyhow.workspace = true
+bincode.workspace = true
+dashmap.workspace = true
+serde.workspace = true
